rename database to supabase to make it easier to find stuff
